Create UI.r
This is the UI.r code for my first R User Interface. Long story short, I saw an attorney through his process of deciding how many stamps or what combination of 49 cents and 21 cents stamps he should post on an envelope that is assigned a certain price. This simple UI was designed to simplify that specific task and save some seconds or minutes every day.